BUG: Trying to remove output directory even if it doesn't exist
When setting 'clean' to 'True' in the configuration file, the output directory is remove to make sure that computation is started from scratch. If the directory does not exist, the script should not try to remove it, or an error would be raised. To avoid this problem, we first check if the directory already exists.

+++ Script/LowRankImageDecomposition.py
@@ -505,7 +505,7 @@ class LowRankImageDecompositionLogic(ScriptedLoadableModuleLogic):
     file_list_file_name = config.file_list_file_name
     im_fns = pyLAR.readTxtIntoList(os.path.join(data_dir, file_list_file_name))
     # 'clean' needs to be done before configuring the logger that creates a file in the output directory
-    if hasattr(config, "clean") and config.clean:
+    if os.path.isdir(result_dir) and hasattr(config, "clean") and config.clean:
         shutil.rmtree(result_dir)
     logger = logging.getLogger(__name__)
     logger.setLevel(logging.INFO)
